Add folder creation functionality
In case specified path to file is located in a sub-folder
of the default directory - create required folder
structure before saving file.
Make sure that specified path is a subdirectory of the default
directory, so that nobody can save files outside the specified
default directory.
'file' form field is now mandatory.