Send entire log to listener only when explicitly asked
Some logs are too large to be sent to the user automatically,
each time he/she starts to listen them.
When user starts to listen to a log file - send contents that
are already present in the file to the user only in case user
specifies the flag "fromStart":true in the request.