Improve log view page
Use all the page width for log entries.
Add a background for a log part of the page.
Change direction of the log, so that it appends new entries
at the bottom which is a more natural log display direction.
Make page scroll sticky, meaning that if user is looking at the bottom
of the page, when a new entry gets added to the log - the page
is scrolled down appropriately. This will only happen if user is looking
at the bottom of the page. Otherwise - there will be no scroll.