Fix line interrupts in FallbackFileWatcher
"tail" binary may emit incomplete lines into it's STDOUT and STDERR.
We should not emit such incomplete lines though. Each time we
receive an incomplete line from "tail" - we wait for a next chunk
of data and append that incomplete line to it.
This way interrupts in "tail" output will not serve as "line breaks"
in the actual FallbackFileWatcher notifications.