Implement support for zero-copy file upload
Add support for requests with content-type:
application/octet-stream. Such requests will be processed
as if their entire body is the uploaded file. Such requests
should have a "name" URL query parameter, that should be the
same as the "file" request body parameter in multipart requests.
In case a file upload request has multipart content type,
the request will be processed as before.